# src/config/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0"
    },
    "host": "ba-api.***.xyz",
    "basePath": "/internal",
    "schemes": ["https", "http"],
    "produces": ["application/json"],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": ["auth"],
                "summary": "Login with FoodGroup credentials and a receive JWT token.",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "200": {
                        "description": "OK"
                    }
                },
                "content": {
                    "application/json": {}
                },
                "parameters": [{
                    "name": "Credentials",
                    "in": "body",
                    "description": "Login credentials",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "required": [
                            "username",
                            "password"
                        ],
                        "properties": {
                            "username": {
                                "type": "string"
                            },
                            "password": {
                                "type": "string"
                            }
                        }
                    }
                }]
            }
        },
        "/auth/register": {
            "post": {
                "tags": ["auth"],
                "summary": "Register with FoodGroup credentials and a receive JWT token.",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "200": {
                        "description": "OK"
                    }
                },
                "content": {
                    "application/json": {}
                },
                "parameters": [{
                    "name": "Credentials",
                    "in": "body",
                    "description": "Login credentials",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "required": [
                            "username",
                            "password",
                            "email"
                        ],
                        "properties": {
                            "username": {
                                "type": "string"
                            },
                            "password": {
                                "type": "string"
                            },
                            "email": {
                                "type": "string"
                            }
                        }
                    }
                }]
            }
        },
        "/auth/change-password": {
            "post": {
                "tags": ["auth"],
                "summary": "Change password with JWT token, previous and new password.",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "204": {
                        "description": "OK"
                    }
                },
                "content": {
                    "application/json": {}
                },
                "parameters": [{
                        "name": "auth",
                        "in": "header",
                        "description": "Authorization token",
                        "required": true
                    },
                    {
                        "name": "Credentials",
                        "in": "body",
                        "description": "Old password to change, new password to set",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "required": [
                                "oldPassword",
                                "newPassword"
                            ],
                            "properties": {
                                "oldPassword": {
                                    "type": "string"
                                },
                                "newPassword": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ]
            }
        },
        "/user/self/": {
            "get": {
                "tags": ["user"],
                "summary": "Get your info",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "parameters": [{
                        "name": "auth",
                        "in": "header",
                        "description": "Authorization token",
                        "required": true
                    }
                ]
            }
        }
    }
}